chore: move common helpers from solidity-test to solidity plugin
